refactor: DefineResultPage에서 재사용되는 부분 분리 (#49)
